% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overbooking.R
\name{ntickets}
\alias{ntickets}
\title{Overbooking ticket calculator}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{the total number of sits on the plain}

\item{gamma}{probability of overbooking}

\item{p}{the probability that a passenger will show}
}
\value{
return a list with the results for the tickets with the initial input parameters. It also plots the objective vs n function used to solve the problem (normal approx and binomial)
}
\description{
Overbooking ticket calculator
}
\examples{
ntickets(N=400,gamma=0.02,p=0.95)

}
